Bug 1832451 - Vendor wgpu changes. r=webgpu-reviewers,jgilbert
Depends on D177827

Differential Revision: https://phabricator.services.mozilla.com/D177829

+++ dom/webgpu/Adapter.cpp
@@ -27,9 +27,9 @@ void AdapterInfo::GetWgpuName(nsString& s) const {
   s = mAboutSupportInfo->name;
 }
 
-size_t AdapterInfo::WgpuVendor() const { return mAboutSupportInfo->vendor; }
+uint32_t AdapterInfo::WgpuVendor() const { return mAboutSupportInfo->vendor; }
 
-size_t AdapterInfo::WgpuDevice() const { return mAboutSupportInfo->device; }
+uint32_t AdapterInfo::WgpuDevice() const { return mAboutSupportInfo->device; }
 
 void AdapterInfo::GetWgpuDeviceType(nsString& s) const {
   switch (mAboutSupportInfo->device_type) {
